2 warning patches Courtesy of Leon Brocard <acme@astray.com>
git-svn-id: https://svn.parrot.org/parrot/trunk@5786 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ pf/pf_items.c
@@ -572,7 +572,7 @@ PF_store_string(opcode_t *cursor, STRING *s)
             }
         }
     }
-    assert( ((int)charcursor & 3) == 0);
+    assert( ((long)charcursor & 3) == 0);
     LVALUE_CAST(char *, cursor) = charcursor;
     return cursor;
 }
